+++ src/wipeout/sfx.c
@@ -227,7 +227,7 @@ void sfx_set_position(sfx_t *sfx, vec3_t pos, vec3_t vel, float volume) {
 	float distance = vec3_len(relative_position);
 
 	sfx->volume = clamp(scale(distance, 512, 32768, 1, 0), 0, 1) * volume;
-	sfx->pan = -sin(atan2(cam->position.x - pos.x, cam->position.z - pos.z) + cam->angle.y);
+	sfx->pan = -sinf(atan2(cam->position.x - pos.x, cam->position.z - pos.z) + cam->angle.y);
 
 	// Doppler effect
 	float away = vec3_dot(relative_velocity, relative_position) / distance;
@@ -366,7 +366,7 @@ void sfx_stero_mix(float *buffer, uint32_t len) {
 			sfx->position += sfx->pitch;
 			if (sfx->position >= source->len) {
 				if (flags_is(sfx->flags, SFX_LOOP)) {
-					sfx->position = fmod(sfx->position, source->len);
+					sfx->position = fmodf(sfx->position, source->len);
 				}
 				else {
 					flags_rm(sfx->flags, SFX_PLAY);
